refactor: extract drift package (CheckDrift and helpers); move shared test helpers to testharness
