+++ internal/testharness/testharness.go
@@ -86,3 +86,45 @@ func AssertFileContains(t *testing.T, dir, rel, substr string) {
 	content := ReadFile(t, dir, rel)
 	require.Contains(t, content, substr, "file %s does not contain %q", rel, substr)
 }
+
+// MinimalUpstream initialises a local upstream git repo with a minimal
+// .gitspork.yml (upstream_owned only, no templated block) and one file.
+// Returns the temp dir and the initial commit hash.
+func MinimalUpstream(t *testing.T) (string, plumbing.Hash) {
+	t.Helper()
+	dir := t.TempDir()
+	repo, err := gogit.PlainInit(dir, false,
+		gogit.WithDefaultBranch(plumbing.NewBranchReferenceName("main")),
+	)
+	require.NoError(t, err)
+	require.NoError(t, os.MkdirAll(filepath.Join(dir, "upstream-owned"), 0755))
+	require.NoError(t, os.WriteFile(filepath.Join(dir, "upstream-owned", "file.txt"), []byte("upstream content\n"), 0644))
+	require.NoError(t, os.WriteFile(filepath.Join(dir, ".gitspork.yml"), []byte("upstream_owned:\n- upstream-owned/**\n"), 0644))
+	hash := CommitAllWithMessage(t, repo, "initial")
+	return dir, hash
+}
+
+// EmptyDownstream initialises a bare local downstream git repo ready for
+// Integrate to write into.
+func EmptyDownstream(t *testing.T) string {
+	t.Helper()
+	dir := t.TempDir()
+	_, err := gogit.PlainInit(dir, false,
+		gogit.WithDefaultBranch(plumbing.NewBranchReferenceName("main")),
+	)
+	require.NoError(t, err)
+	return dir
+}
+
+// CommitAllWithMessage stages and commits all changes in repo, returning the
+// resulting commit hash.
+func CommitAllWithMessage(t *testing.T, repo *gogit.Repository, message string) plumbing.Hash {
+	t.Helper()
+	wt, err := repo.Worktree()
+	require.NoError(t, err)
+	require.NoError(t, wt.AddWithOptions(&gogit.AddOptions{All: true}))
+	sig := &object.Signature{Name: "gitspork-test", Email: "gitspork-test@localhost", When: time.Now()}
+	hash, err := wt.Commit(message, &gogit.CommitOptions{Author: sig})
+	require.NoError(t, err)
+	return hash
+}
